requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
pandas>=1.3.0
tqdm>=4.60.0
Pillow>=8.2.0
opencv-python>=4.5.0
numpy>=1.21.0
kaggle>=1.5.12
wandb >=0.23.1